tRevert "x"
This reverts commit 58575f7e292545d71aa9ec456437cfe7e2832bf8.

+++ modules/businessproject/businessproject-api/src/main/java/com/bjike/goddess/businessproject/excel/BusinessContractExport.java
@@ -197,7 +197,7 @@ public class BusinessContractExport extends BaseTO {
      * 合同规模数量
      */
     @ExcelHeader(name="合同规模数量")
-    private Integer scaleContract;
+    private Double scaleContract;
 
     /**
      * 合同规模数是否有差异
@@ -704,11 +704,11 @@ public class BusinessContractExport extends BaseTO {
         this.scale = scale;
     }
 
-    public Integer getScaleContract() {
+    public Double getScaleContract() {
         return scaleContract;
     }
 
-    public void setScaleContract(Integer scaleContract) {
+    public void setScaleContract(Double scaleContract) {
         this.scaleContract = scaleContract;
     }
