+++ modules/businessproject/businessproject-api/src/main/java/com/bjike/goddess/businessproject/excel/BusinessContractTemplateExcel.java
@@ -199,7 +199,7 @@ public class BusinessContractTemplateExcel extends BaseTO {
      * 合同规模数量
      */
     @ExcelHeader(name="合同规模数量")
-    private Integer scaleContract;
+    private Double scaleContract;
 
     /**
      * 合同规模数是否有差异
@@ -691,11 +691,11 @@ public class BusinessContractTemplateExcel extends BaseTO {
         this.scale = scale;
     }
 
-    public Integer getScaleContract() {
+    public Double getScaleContract() {
         return scaleContract;
     }
 
-    public void setScaleContract(Integer scaleContract) {
+    public void setScaleContract(Double scaleContract) {
         this.scaleContract = scaleContract;
     }
